New Contorllers Added

+++ Application/ServiceProviders/Commands/DeleteServiceProvider/DeleteServiceProviderCommandHandler.cs
@@ -1,5 +1,5 @@
 ﻿using BookingServices.Application.Common.Interfaces;
-using BookingServices.Application.ServiceProviders.Queries.GetServiceProviderDetail;
+using BookingServices.Domain.ValueObjects;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -9,21 +9,31 @@ namespace BookingServices.Application.ServiceProviders.Commands.DeleteServicePro
 {
     public class DeleteServiceProviderCommandHandler : IRequestHandler<DeleteServiceProviderCommand>
     {
+
         private readonly IBookingServicesDbContext _context;
         public DeleteServiceProviderCommandHandler(IBookingServicesDbContext bookingServicesDbContext)
         {
             _context = bookingServicesDbContext;
-
         }
 
         public async Task Handle(DeleteServiceProviderCommand request, CancellationToken cancellationToken)
         {
-            var serviceProvider = await _context.ServiceProviders.Where(s => s.Id == request.ServiceProviderId )
-               .FirstOrDefaultAsync(cancellationToken);
+            var serviceProvider = await _context.ServiceProviders
+               .Where(s => s.Id == request.Id && s.IsActive == true).FirstOrDefaultAsync(cancellationToken);
 
-            _context.ServiceProviders.Remove(serviceProvider);
-            await _context.SaveChangesAsync(cancellationToken);
+            if (serviceProvider != null)
+            {
+                serviceProvider.Address = new Address(
+                    serviceProvider.Address.Street,
+                    serviceProvider.Address.City,
+                    serviceProvider.Address.ZipCode,
+                    serviceProvider.Address.Number);
+                serviceProvider.ContactPerson = new PersonName(serviceProvider.ContactPerson.FirstName, serviceProvider.ContactPerson.LastName);
+                serviceProvider.Email = new Email(serviceProvider.Email.UserName, serviceProvider.Email.DomainName);
 
+                _context.ServiceProviders.Remove(serviceProvider);
+                await _context.SaveChangesAsync(cancellationToken);
+            }
         }
     }
 }
